[정인호] refactor: utils 패키지에 공통 클래스 통합
